LPS-66791 Move to internal package
